"ads2_m12_a1_1: 0/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS-2-assignments/m12/ADS-2_Exam-2/Solution.java
@@ -30,7 +30,7 @@ public class Solution {
 			String[] distPath = sc.nextLine().split(" ");
 			Edge e = new Edge(Integer.parseInt(distPath[0]),
 			                  Integer.parseInt(distPath[0]),
-			                  Integer.parseInt(distPath[0]));
+			                  Double.parseDouble(distPath[0]));
 			graph.addEdge(e);
 		}
 		String caseToGo = sc.nextLine();
